+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] 11-06-2023
+
+- Added duplicate testing.
+
+
 ## [0.1.0] 09-06-2023
 
 ### Added
 
-- Initial version containing the base, changelog, and export modules
+- Initial version containing the base, changelog, and export modules.
 - Automated validation using the MIBiG JSON schema.